AI Agent/pyproject.toml
[project]
name = "ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-mcp-adapters>=0.1.8",
    "langchain-openai>=0.3.27",
    "langgraph>=0.5.1",
    "python-dotenv>=1.1.1",
]
